+++ src/main.zig
@@ -1,9 +1,12 @@
 const std = @import("std");
 
 const jsonic = @import("jsonic");
+const StaticJSON = jsonic.StaticJSON;
+const DynamicJSON = jsonic.DynamicJSON;
+
 
 pub fn main() !void {
-    std.debug.print("Hello, World!\n", .{});
+    std.debug.print("Code coverage examples - \n", .{});
 
     // Let's start from here...
 
@@ -11,6 +14,7 @@ pub fn main() !void {
     defer std.debug.assert(gpa_mem.deinit() == .ok);
     const heap = gpa_mem.allocator();
 
+    // Static JSON
     {
         const User = struct { name: []const u8, age: u8 };
 
@@ -19,52 +23,62 @@ pub fn main() !void {
         defer heap.free(src);
         std.mem.copyForwards(u8, src, static_str);
 
-        const data = try jsonic.StaticJson.parse(User, heap, src);
+        const data = try StaticJSON.parse(User, heap, src);
         std.debug.print(
-            "Structure Data - name: {s} age: {d}\n", .{data.name, data.age}
+            "Static JSON Parsed Output - name: {s}, age: {d}\n",
+            .{data.name, data.age}
         );
 
-        const json_str = try jsonic.StaticJson.stringify(heap, data);
+        const json_str = try StaticJSON.stringify(heap, data);
         defer heap.free(json_str);
 
-        std.debug.print("Stringify Data - {s}\n", .{json_str});
+        std.debug.print(
+            "Stringified JSON from Zig Structure - {s}\n", .{json_str}
+        );
         try jsonic.free(heap, data);
     }
 
+    // Dynamic JSON - Array
     {
         const static_str = "[\"John Doe\", 40]";
         const src = try heap.alloc(u8, static_str.len);
         defer heap.free(src);
         std.mem.copyForwards(u8, src, static_str);
 
-        var dyn_json = try jsonic.DynamicJson.init(heap, src, .{});
+        var dyn_json = try DynamicJSON.init(heap, src, .{});
         defer dyn_json.deinit();
 
         const json_data = dyn_json.data().array;
         const item_1 = json_data.items[0].string;
         const item_2 = json_data.items[1].integer;
-        std.debug.print("Array Item - Name: {s} Age: {}\n", .{item_1, item_2});
+        std.debug.print(
+            "JSON Array Items - Name: {s}, Age: {}\n",
+            .{item_1, item_2}
+        );
     }
 
+    // Dynamic JSON - String Array
     {
-        const SliceType = []const []const u8;
+        const Str = []const u8;
+        const StrArray = []const Str;
         const static_str = "[\"John Doe\", \"Jane Doe\"]";
         const src = try heap.alloc(u8, static_str.len);
         defer heap.free(src);
         std.mem.copyForwards(u8, src, static_str);
 
-        var dyn_json = try jsonic.DynamicJson.init(heap, src, .{});
+        var dyn_json = try DynamicJSON.init(heap, src, .{});
         defer dyn_json.deinit();
 
         const value = dyn_json.data();
-        const result = try jsonic.DynamicJson.parseInto(SliceType, heap, value, .{});
-        const str = try jsonic.StaticJson.stringify(heap, result);
+        const result = try DynamicJSON.parseInto(StrArray, heap, value, .{});
+        const str = try StaticJSON.stringify(heap, result);
         defer heap.free(str);
 
-        std.debug.print("Stringify Result:\n{s}\n", .{str});
+        std.debug.print("Stringified Array Items:\n{s}\n", .{str});
         try jsonic.free(heap, result);
     }
 
+    // Dynamic JSON - Object
     {
         const static_str =
         \\ {
@@ -82,17 +96,18 @@ pub fn main() !void {
         std.mem.copyForwards(u8, static_input, static_str);
         defer heap.free(static_input);
 
-        var json_value = try jsonic.DynamicJson.init(heap, static_input, .{});
+        var json_value = try DynamicJSON.init(heap, static_input, .{});
         defer json_value.deinit();
 
         const value = json_value.data().object;
         const joy = value.get("feelings").?.object.get("joy").?.integer;
-        std.debug.print("Joy: {d}\t", .{joy});
+        std.debug.print("JSON Object - Joy: {d}\t", .{joy});
 
         const hobby = value.get("hobby").?.array.items[1].string;
-        std.debug.print("Hobby: {s}\n\n", .{hobby});
+        std.debug.print("JSON Object - Hobby: {s}\n", .{hobby});
     }
 
+    // Dynamic JSON - Mixed
     {
         const Feelings = struct { fear: f64, joy: i32 };
         const Foo = enum {Bar, Baz};
@@ -121,17 +136,17 @@ pub fn main() !void {
         std.mem.copyForwards(u8, static_input, static_str);
         defer heap.free(static_input);
 
-        var json_value = try jsonic.DynamicJson.init(heap, static_input, .{});
+        var json_value = try DynamicJSON.init(heap, static_input, .{});
         defer json_value.deinit();
 
         const src = json_value.data();
 
-        const result = try jsonic.DynamicJson.parseInto(User, heap, src, .{});
-        std.debug.print("result {any}\n", .{result});
-        const str = try jsonic.StaticJson.stringify(heap, result);
+        const result = try DynamicJSON.parseInto(User, heap, src, .{});
+        std.debug.print("Mixed JSON Result {any}\n", .{result});
+        const str = try StaticJSON.stringify(heap, result);
         defer heap.free(str);
 
-        std.debug.print("Stringify Result:\n{s}\n", .{str});
+        std.debug.print("Stringified JSON Result {s}\n", .{str});
         try jsonic.free(heap, result);
     }
 }