+++ src/root.zig
@@ -4,5 +4,5 @@
 const parser = @import("./core/parser.zig");
 
 pub const free = parser.free;
-pub const StaticJson = parser.Static;
-pub const DynamicJson = parser.Dynamic;
+pub const StaticJSON = parser.Static;
+pub const DynamicJSON = parser.Dynamic;